uint32n

+++ common.go
@@ -55,11 +55,8 @@ func cryptorandUint32() (uint32, error) {
 	return binary.BigEndian.Uint32(buf[:]), nil
 }
 
-// uint32n returns, as an int32, a non-negative pseudo-random number in [0,n).
-// n must be > 0, but int31n does not check this; the caller must ensure it.
-// int31n exists because Int31n is inefficient, but Go 1 compatibility
-// requires that the stream of values produced by math/rand remain unchanged.
-// int31n can thus only be used internally, by newly introduced APIs.
+// uint32n returns, as a uint32, a non-negative pseudo-random number in [0,n).
+// It is adapted from math/rand.int31n.
 //
 // For implementation details, see:
 // https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction
@@ -71,8 +68,8 @@ func uint32n(n uint32) (uint32, error) {
 	}
 	prod := uint64(v) * uint64(n)
 	low := uint32(prod)
-	if low < uint32(n) {
-		thresh := uint32(-n) % uint32(n)
+	if low < n {
+		thresh := -n % n
 		for low < thresh {
 			v, err = cryptorandUint32()
 			if err != nil {
